fix issues with postgres

+++ src/model/user.rs
@@ -5,7 +5,7 @@ use tokio_pg_mapper_derive::PostgresMapper;
 #[derive(Deserialize, PostgresMapper, Serialize)]
 #[pg_mapper(table = "users")] // singular 'user' is a keyword..
 pub struct User {
-    pub id: u32,
+    pub id: i64,
     pub email: String,
     pub username: String,
     pub role: String,
@@ -15,6 +15,6 @@ pub struct User {
 
 #[derive(Serialize, Deserialize, Clone, FromRequest)]
 pub struct UserClaims {
-    pub id: u32,
+    pub id: i64,
     pub role: String,
 }
